Move json extension to the same file as JSON type protocol
