+++ SwiftyJSONModel/JSONTypeProtocol.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import SwiftyJSON
 
 public protocol JSONType {
     var bool: Bool? { get }
@@ -61,4 +62,13 @@ public extension JSONType {
         guard let dictValue = dictionary else { throw JSONModelError.invalidElement }
         return dictValue
     }
+}
+
+extension JSON: JSONType {
+    public init(bool: Bool) { self.init(bool) }
+    public init(int: Int) { self.init(int) }
+    public init(double: Double) { self.init(double) }
+    public init(string: String) { self.init(string) }
+    public init(array: [JSON]) { self.init(array) }
+    public init(dictionary: [String: JSON]) { self.init(dictionary) }
 }
